[Compute]Check StatefulSet ObservedGeneration
To consider the compute service deployment ready the controller needs
to check not just the number of ready replicas of the statefulset but
also that the staterfulset controller saw the last update from the
compute controller. So the check is changed to compare the
Generation of the StatefulSet Spec and the ObservedGeneration in its
Status.