Delete MariaDBAccount and Database
Both CR is created by the nova controller for the given cell, so when
that cell is deleted the CRs also needs to be deleted. We also need to
delete the secret connected to the MariaDBAccount as mariadb-operator
does not delete it today.

I also added graceful handling of NotFound during resource deletion.