Add Notfications message bus instance for Nova CRD
Extend CRDs to support a RabbitMqCluster reference as AMQP instance for
notifications.

To provide a way for human user to define the notifications bus for
Nova CR via the OpenStackControlPlane CR's nova section, add a new
field to the NovaSpecCore as that is the struct used when the
OpenStackControlPlaneSpec struct is defined.

At NovaSpec level having just "" or a name of the bus is OK as either
we need a bus, or we don't need notifications at all.  However, for
empedded NovaSpecCore we might need the 3rd possibility to "inherit
from somewhere higher", which we represent as: *string field where null
means inherit, "" means disable notification, "foo" means send
notification to bus foo.

Note that overriding NotificationsBusInstance for cells is not
supported as that implies that consumers of notifications can maintain
simultanous AMQP connections via multiple transport_url endpoints,
which is not a thing in Openstack.

Assume that in the implementation the required information will be
transferred via the existing internal secret via notifications specific
transport_url key (as the transport_url key is already taken by the RPC
bus value). Thus we do not modify other Nova services' CRDs.

Note that the pointer type is redundant for NovaSpec, but since
we have already defined it as a such in NovaSpecCore, and
changing the types system is not planned (requires changes in openstack
operator), it should be used as is for also NovaSpec.
However there shall be only a real use of a nil value in
NovaSpecCore (while unused in NovaSpec).

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>